+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 #### Unreleased changes
 * Nothing yet...
 * Slideshow more responsive on pressing next
+* Slideshow timer resets when next is pressed
 
 #### Version 1.0 alpha 2 - 10/15/2019
 * Fix installer running in wrong mode